Add get_number() to TodoList.
This allows the canonical position of the Task in the list to be
obtained by the TaskId.

This is not the most efficient algorithm. It's O(n) to find the
number of a single Task, which means a straightforward rendering
of an entire TodoList is going to be O(n^2). It might be a good
idea to have methods like incomplete_tasks() that iterate TaskIds
include the task number.