Reimplement TodoList using Dag.
This lays the foundation of introducing dependency relationships
between tasks.

Currently, there are two special root nodes that are used to query
whether a task is complete or incomplete. Incomplete tasks are
parents of the terminal "incomplete_root" and complete tasks are
parents of the terminal "complete_root".

The underlying graph implementation unfortunately yields neighbors
in the reverse order that they were added, which requires
revsersing iterators inefficiently to maintain the ordering
contract that we want for tasks. Tasks should appear in the order
they were added. This may complicate future efforts to adopt a
canonical ordering once dependency relationships between tasks are
allowed, but we'll see.