Add a test for total ordering.
This test doesn't yet pass, but shows the requirement that should
be imposed on the ordering of incomplete tasks: that a blocked
task should show up after all unblocked tasks.

In general, we consider the "depth" of a task to be the maximum
number of edges it can travel to reach the incomplete node. A task
with a given depth must appear after all tasks that have a lower
depth. Tasks with equal depth should appear in the order they were
inserted. This defines a total ordering that promotes the
stability of the positioning of tasks in the list.