Rewrite TodoList implementation.
Instead of trying to do everything with the graph structure, only
use the graph for storing tasks and the dependency relationships
between them. To get a canonical ordering for incomplete tasks,
we introduce a helper struct, Layering, which keeps track of the
insertion order and "depth" of tasks, where depth is defined as
the maximum number of edges one needs to traverse to reach an
unblocked task. This allows us to ensure that every task that has
a certain depth will appear earlier in the list than every task
that has a higher depth, and fairly straightforwardly provides a
cheap way to iterate the tasks within a depth layer by the order
of insertion.