Add tests for some behavior for 'check' and 'restore'.
Ideally, these should be symmetric. If you try to complete a task
with incomplete dependencies, or restore a task with complete
antidependencies, it should be an error. We can have a --force
flag for these commands that make them recursive, so that
completing a task with incomplete dependencies will implicitly
complete those dependencies, recursively; and likewise, restoring
a task with complete antidependencies will restore those
antidependencies, recursively.

Currenlty, 'restore' has the --force behavior and 'check' has the
non---force behavior.