Use TaskSet instead of HashSet for deps/adeps.
This allows the tasks to be iterated in sorted order, if desired.
This is currently an O(n^2*log(n)) operation for n tasks in the
set, so doesn't scale very well.

There are two potential other algorithms that might be faster:

* O(n), for n as the size of the entire to-do list, if we iterate
  all_tasks() and filter by presence in a HashSet.
* O(n*l*log(n)), where l is the maximum size of the layer any of
  the tasks is in, if we use the depth hints in Layering to find
  the position more efficiently.

In the meantime, this algorithm works and the implementation is
opaque to the end user.