Use a text editor in 'edit' command.
If no --desc arg is provided, then show a text editor that lets
you edit the descriptions of tasks interactively. The text editor
will show the task numbers and descriptions of the tasks that
match the given keys. Any changes the user makes to the task
descriptions before saving will be committed to the todo list and
the tasks will be printed, confirming the changes.

Errors will be printed if the user messes up the number or
delimiter between the number and the description in the text
editor.