Show completion dates in log.
Introduce a trait for controlling the "current time" for testing,
inject that into app::todo, and introduce a new Fixture class in
app::testing to allow easier manipulation of the environment (and
avoid as much syntax bloat when making argument lists, since we
can pass the command line arguments as one string instead of a
slice of strings).

The log implementation will show the date YYYY-MM-DD for when the
tasks were completed. Repeat dates are not shown, i.e. a task
will not show its completion date if a task above has the same
completion date.