Sort tasks by implicit priority.
If b is blocked by a, then a inherits b's priority. The implicit
priority of a task is, generally, the maximum of its own priority
and the implicit priority of its adeps.

This makes sense because if you block a high-priority task on a
low-priority one, that blocking task should become high-priority.