Make set ops for TaskSet owning.
This avoids unnecessary copies in a lot of cases.
However, in some cases, such as the uses in 'merge', we have to
clone() a TaskSet multiple times. However, it's more common to
repeatedly accumulate small sets into a larger set using fold(),
which benefits from re-using one of the two HashSets from the
source TaskSets.