Add 'bottom' commnd.
This is the opposite of the 'top' command: if no keys are given, it
prints tasks that have no dependencies, and if keys are given, it prints
the direct anti-dependencies of the given tasks.

The no-key case is useful for finding tasks that can be run in parallel,
or tasks that "start" a project, or tasks should be put into a structure
with other bottom tasks.

The with-key case is useful for scripting: you can match the tasks that
are blocked by a given task with $(todo bottom x).