Record PrintableTasks directly in tests.
This avoids unnecessary copies in tests and avoids the need to separate
a "PrintedTaskInfo" struct from a "PrintableTask" struct.

This also straightens out the lifetime invariants of TodoPrinter.

We weren't comparing deps_stats and adeps_stats before, but now that
we're comparing PrintableTask objects directly, we have to specify the
full info in tests. This is why most of this change is updating tests.