Allow 'find' to search for inexact tag matches.
Before, to search tags, you had to explicitly pass the --tag flag.
In addition, matches had to be exact (you couldn't search for a
substring of a tag).

Now, all tasks with implicit tags that partially match the search terms
will be included in results. Tasks with descriptions that match the
search terms will be highlighted with a * selector, helping the user
tell the difference from explicit and implicit matches.