Use checked_ilog10() for calculating digit width.
The unchecked version would crash on empty data models, because the
log base 10 of 0 is undefined. Bug the point is to calculate the digit
width so if there are no tasks to display it doesn't really matter what
the digit width is. I chose to default to 1 because it's possible the
digit width could affect the formatting of non-task printable info at
some point.