Put max_index_digits() logic to PrintableResult.
There was a subtle bug in the previous implementation which was not
covered by tests: if the number of digits of a new task was greater than
the max number of digits in the list before the new task was added, the
max number of digits would be wrong, and the printed tasks could be
misaligned.

By putting the method to calculate the max number of digits in the app
result we force the calculation to be made after mutating the list.

A side effect of this is that we are also forced to instantiate the
printer only after making a transaction with the data model. This is
probably fine but will make it harder to ever stream an app result to
a printer if I ever want to do that.