Display tags in reverse sorted order.
This allows more "general" tags to appear first, and work from left to
right progressing from "general" to "specific". This also has
higher-priority tags appear closer to the right, closer to the
description. Furthermore, this leaves the order that the tags are
displayed independent of the order in which the tags were applied, so
adjacent tasks with the same tags will show their tags in the same
order.