Snoozed tasks are not put in a higher layer.
Before, a snoozed task whose natural depth was 0 (i.e. it had no
incomplete dependencies) would be put in layer 1, as if it had
incomplete dependencies after all. This was meant to sort them after
unblocked incomplete tasks but before blocked tasks on layer 1.

However, it was possible for a snoozed task to be sorted after actually
blocked tasks, including tasks that the snoozed task was blocking, which
would violate the intended invariant that tasks would always be sorted
after their dependencies.

The new approach uses whether a task is snoozed to determine the first
comparison when sorting tasks, so that a snoozed task always appears
after all unsnoozed tasks on the same layer. This generalizes snoozed
tasks to layers other than 0 and 1, which can be useful if you want to
snooze a blocked task so that it won't appear right away when you
complete its dependencies.

This should be a fully backwards-compatible change: all tasks are
considered for unsnoozing no matter what layer they are on, and once
their start date passes, they will be put on the correct layer. No
fields have been added or removed from the data model. We use the fact
of whether the start date is the same as the creation date to determine
whether the task is snoozed in order to avoid having to add a new field
to the data model.