Add a clean subcommand to the CLI.
This is used to clean up corrupt or invalid data in the database.
In particular, old versions of the app could get the database into a
state that violates layering rules, sorting rules, or stale id rules.

This implementation isn't guaranteed to fix all of those problems, but
at the very least it does a simple brute force approach of updating the
depth and position of all incomplete tasks, which seems to be enough to
fix the problems I am seeing so far.