Move builtin apps to builtin_apps crate.
Also move Application trait definition to app crate.

Most surprising source of pain was getting the boolean for whether the
app modified the list to be an enum, just because there were so many
compiler errors in the tests that got queued up that rust-analyzer got
overwhelmed even on gitpod.