Truncate tags if they don't fit on one line.
Truncation is done by greedily adding tags from the front and back of
the list of tags until they meet in the middle at the capacity of the
line, including space for the truncation delimiter string. The algorithm
tries to balance the left and right sides of the truncation point by
always choosing the tag that makes its respective side of the truncation
point shortest when added.

There are a surprising number of edge cases and errors that need to be
considered, so there are unit tests in truncate_test.rs.

I should make sure to also add tests for simple_todo_printer.rs printing
truncateed tags.

I think it's likely that I'll want to toggle this behavior on and off
in certain circumstances. For example, maybe tasks that are matches for
a 'todo get' query should not truncate their tags so the user can see
the full set of tags when inspecting a task directly. Or the user may
wish to use the config file to disable this behavior entirely, or change
the delimiter character.