Do not truncate tags that match 'todo get' queries.
By default, implicit tag sets are truncated in the middle if they don't
all fit on one line, but you can use 'todo get' to see the full tag set
for the task you're "getting".

This adds a new field to PrintableTask to opt into or out of truncation,
and updates the SimpleTodoPrinter to respect this field. Additionally,
the formatting utility functions opt into truncation by default and the
'todo get' implementation and tests are adjusted to opt out when the
task being printed is an exact match for the query.