Don't include complete tasks on ambiguous 'new' queries.
If you create a task that is blocked on, blocked by, or put before,
after, or by another task that is looked up by an ambiguous query,
the query will resolve to only the incomplete tasks that match the query
if there are both complete and incomplete matches. This matches the
behavior of `get` without the `-d` flag.

There is no way to, with one `new` command, create a task that e.g.
blocks both incomplete and complete tasks with the same name except by
passing the positional indices of those tasks, or using scripting like
'todo new a -b $(todo get b -d)', which will block a on both incomplete
and complete tasks with the name 'b'.