Snoozing a task until before its snooze date is a no-op.
Most of the time snoozing an already-snoozed task happens as a result of
a scripting query like 'todo snooze $(todo get work) --until tomorrow'
where some of the tasks that match the 'get' query are already snoozed.

If the new snooze date overrides the old one in cases like this, you can
get bombarded with more unsnoozed tasks than expected when the snooze
date arrives.

Now, if the new snooze date is before the old one, the snooze operation
is a no-op, but a warning is printed to let the user know that the task
was already snoozed until a later date. If this warning message becomes
too annoying I may change or remove it.