Fix clippy lint in runner crate.
The new lint gives a warning when a field is never read outside
a Debug derive, which doesn't count. This patch fixes it by
explicitly extracting the inner field from the error struct.

This also adds the command that failed to spawn a paginator to
the error message to help the user debug the problem if they
misconfigured the paginator.