Cache the task status in a persistable field.
The status() function that computed the status by first checking
for presence in the completion list was an O(n) lookup which was
a bottleneck when running `todo` or `todo -a` or `todo log` which
would essentially run the lookup on all tasks.

This change introduces a cached status field that is updated when
the status changes (tasks become blocked/unblocked/complete/etc)
and short-circuits this to an O(1) lookup.